fix: avoid truncating the account order history
