+++ src/hooks/query-options.ts
@@ -3,11 +3,10 @@ import { getCategoryById } from "@/lib/category-utils";
 import {
   getBestSellingProducts,
   getDashboardMetrics,
-  getOrders,
   getRevenueChartData,
 } from "@/lib/dashboard-utils";
 import { getCheckoutOrder } from "@/lib/order-storage";
-import { getOrderById } from "@/lib/order-utils";
+import { getOrderById, getOrders } from "@/lib/order-utils";
 import {
   getAdminProducts,
   getProductById,
